Update create_checkerboard.py - update floats to ints
Define ints instead of floats for h/w/size

+++ backend/src/packages/chaiNNer_standard/image/create_images/create_checkerboard.py
@@ -37,11 +37,11 @@ from .. import create_images_group
     ],
 )
 def create_checkerboard_node(
-    width: float,
-    height: float,
+    width: int,
+    height: int,
     color_1: Color,
     color_2: Color,
-    square_size: float,
+    square_size: int,
 ) -> np.ndarray:
     img = np.zeros(
         (height, width, 3), dtype=np.float32
